Actual login (#29)
* Minimal version of django auth backend

Replacement of #22. We only really want the django side to have the minimal
amount of info from the Tyler API (we'll cache the call to firm info / user info, sure,
but it doesn't need to be a part of the database model).

Mostly just copied over `authenticate.py` (the parts that weren't hallucinated), and
some of `models.py` and `settings_base.py`. `auth_views.py` likely will need some more
changes, but they'll have to come with bigger refactors / dedup of code I think.

* Fix tests

* When not logged in, redirect to login page

Fixes most of #14.